modified reward logic to prevent agent from being rewarded for moving away from the target delivery location
During testing, I observed that the agent was getting high rewards for exploring the environment, even when its actions led it far from the delivery location.

This happened because I was rewarding the agent anytime the sensors were not in contact with the border or an obstacle.

I removed the rewards from the safety function and now only punish the agent when the sensors collide with an obstacle.